Add the Fortran BMI spec file snow model BMI directory
Placing "bmi.f90" alongside the BMI for the snow model greatly
simplifies the build process (removing the "bmi" directory entirely),
plus it doesn't build a shared object for the BMI spec file, which
could conflict if installed in a location with other BMI-ed Fortran
models. The BMI spec is "baked in" to the BMI implementation of the
snow model.